Moar accurate classes
