+++ src/main/java/com/vandendaelen/openbackup/helpers/FileHelper.java
@@ -1,0 +1,18 @@
+package com.vandendaelen.openbackup.helpers;
+
+import com.vandendaelen.openbackup.OpenBackup;
+import org.apache.commons.io.FileUtils;
+
+import java.io.File;
+
+public class FileHelper {
+
+    public static void deleteWorldForlder(File worldFolder){
+        try{
+            FileUtils.deleteDirectory(worldFolder);
+        }
+        catch (Exception e) {
+            OpenBackup.logger.info(e.getMessage());
+        }
+    }
+}
